+++ src/app/store.js
@@ -1,8 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit';
-import counterReducer from '../features/counter/counterSlice';
+import greetingReducer from '../features/greeting/greetingSlice';
 
 export const store = configureStore({
   reducer: {
-    counter: counterReducer,
+    message: greetingReducer,
   },
 });
+
+export default store;
