+++ src/features/greeting/greetingSlice.js
@@ -1,0 +1,37 @@
+import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
+import axios from 'axios';
+
+export const fetchGreetings = createAsyncThunk(
+  'greeting/fetchGreetings',
+  async () => {
+    const response = await axios.get('/api/greetings');
+    return response.data;
+  },
+);
+
+const initialState = {
+  message: '',
+};
+
+export const counterSlice = createSlice({
+  name: 'greeting',
+  initialState,
+  reducers: {},
+  extraReducers: (builder) => {
+    builder
+      .addCase(fetchGreetings.pending, (state) => {
+        // eslint-disable-next-line no-param-reassign
+        state.status = 'loading';
+      })
+      .addCase(fetchGreetings.fulfilled, (state, action) => {
+        // eslint-disable-next-line no-param-reassign
+        state.status = 'idle';
+        // eslint-disable-next-line no-param-reassign
+        state.message = action.payload;
+      });
+  },
+});
+
+export const { increment, decrement, incrementByAmount } = counterSlice.actions;
+
+export default counterSlice.reducer;
